// client/src/app/api/openapi/backend/model/dtoEventInventory.ts
/**
 * Messe Server API
 * API für die Messe-Anwendung
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DtoEventInventory { 
    id?: number | null;
    startedAt?: string | null;
    tradeEventId?: number | null;
}