// client/src/app/api/openapi/backend/model/dtoTradeEvent.ts
/**
 * Messe Server API
 * API für die Messe-Anwendung
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DtoTradeEvent { 
    id?: number | null;
    name?: string | null;
}